some restructuring to fit with some of the suggestions
